+++ src/extra/Item.jsx
@@ -1,7 +1,7 @@
 import React, { Fragment, useState, useRef } from "react";
 import { useDrag, useDrop } from "react-dnd";
 import Window from "./Window.jsx";
-import ITEM_TYPE from "../data/types.js";
+import ITEM_TYPE from "../../data/types.js";
 
 const Item = ({ item, index, moveItem, status }) => {
     //ref allows us to imperatively modify a child
